Javadoc Comments and turned floats into doubles

+++ a1/AllTest.java
@@ -30,16 +30,16 @@ public class AllTest {
 	public void testScale(){
 		//Identity
 		Vector3D a = new Vector3D(1.0, 1.0, 1.0);
-		Vector3D idAResult = a.scale(1);
+		Vector3D idAResult = a.scale(1.0);
 		Vector3D idAExpected = new Vector3D(1.0, 1.0, 1.0);
 		assertTrue(idAResult.equals(idAExpected));
 		//Zero
 		Vector3D b = new Vector3D(3.21,4.1,5.0);
-		Vector3D zeroBResult = b.scale(0);
+		Vector3D zeroBResult = b.scale(0.0);
 		Vector3D zeroBExpected = new Vector3D(0.0, 0.0, 0.0);
 		assertTrue(zeroBResult.equals(zeroBExpected));
 		//General Case
-		float scalar = 2;
+		double scalar = 2;
 		Vector3D c = new Vector3D(5.2, 3.3, 4.21);
 		Vector3D gcCResult = c.scale(scalar);
 		Vector3D gcCExpected = new Vector3D(10.4, 6.6, 8.42);
